+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/HomeDialogConfirmButton.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.home.presentation.component.dialog
+package com.teamoffroad.feature.home.presentation.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.interaction.MutableInteractionSource
