+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/HomeQuestContents.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.home.presentation.component.quest
+package com.teamoffroad.feature.home.presentation.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxWidth
