+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/HomeQuestTitle.kt
@@ -1,4 +1,4 @@
-package com.teamoffroad.feature.home.presentation.component.quest
+package com.teamoffroad.feature.home.presentation.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Text
