+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/HomeSelectedEmblemState.kt
@@ -1,0 +1,7 @@
+package com.teamoffroad.feature.home.presentation.component
+
+enum class HomeSelectedEmblemState {
+    SAME_WITH_ORIGIN_EMBLEM,
+    SAME_WITH_SELECTED_CHANGE_EMBLEM,
+    SELECTED_CHANGE_EMBLEM
+}
