--- feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/dialog/SelectedEmblemState.kt
@@ -1,7 +1,0 @@
-package com.teamoffroad.feature.home.presentation.component.dialog
-
-enum class SelectedEmblemState {
-    SAME_WITH_ORIGIN_EMBLEM,
-    SAME_WITH_SELECTED_CHANGE_EMBLEM,
-    SELECTED_CHANGE_EMBLEM
-}
